+++ client/src/components/overview/ZoomedPreview.jsx
